implement workflow backend

+++ apps/api/src/api/data-stores/records/controller.ts
@@ -1,12 +1,12 @@
-import { dataRecordsModel, dataStoresModel } from "../../../models";
 import fs from 'fs-extra';
 import path from 'path';
 import config from "../../../config";
 import { RecordsApi } from "./types";
 import { HttpError, HttpCode } from "@pestras/backend/util";
-import { DataStore, Field } from "@pestras/shared/data-model";
+import { DataRecordState, DataStore, Field } from "@pestras/shared/data-model";
 import { NextFunction } from 'express';
 import { Serial } from "@pestras/shared/util";
+import { dataRecordsModel, dataStoresModel } from '@pestras/backend/models';
 
 export const controller = {
 
@@ -15,7 +15,7 @@ export const controller = {
   async search(req: RecordsApi.SearchReq, res: RecordsApi.SearchRes, next: NextFunction) {
     try {
       res.json(await dataRecordsModel.search(req.params.serial, req.body));
-      
+
     } catch (error) {
       next(error);
     }
@@ -23,8 +23,8 @@ export const controller = {
 
   async getBySerial(req: RecordsApi.GetBySerialReq, res: RecordsApi.GetBySerialRes, next: NextFunction) {
     try {
-      res.json(await dataRecordsModel.getBySerial(req.params.serial, req.params.record));
-      
+      res.json(await dataRecordsModel.getBySerial(req.params.serial, req.params.record, req.params.state));
+
     } catch (error) {
       next(error);
     }
@@ -33,7 +33,7 @@ export const controller = {
   async getHistory(req: RecordsApi.GetHistoryReq, res: RecordsApi.GetHistoryRes, next: NextFunction) {
     try {
       res.json(await dataRecordsModel.getHistory(req.params.serial, req.params.record));
-      
+
     } catch (error) {
       next(error);
     }
@@ -43,14 +43,9 @@ export const controller = {
   // -------------------------------------------------------------------------------------------------
   async create(req: RecordsApi.CreateReq, res: RecordsApi.CreateRes, next: NextFunction) {
     try {
-      const ds: DataStore | null = await dataStoresModel.getBySerial(req.params.serial, { type: 1, settings: 1, fields: 1, state: 1 });
-      
-      if (!ds)
-        throw new HttpError(HttpCode.NOT_FOUND, 'dataStoreNotFound');
-
       const serial = Serial.gen("RCD");
       const files = req.files as Express.Multer.File[] | undefined;
-  
+
       if (files && files.length) {
         const paths: Record<string, string> = {};
 
@@ -67,8 +62,8 @@ export const controller = {
 
         Object.assign(req.body, paths);
       }
-  
-      res.json(await dataRecordsModel.create(ds, serial, req.body, res.locals.issuer));
+
+      res.json(await dataRecordsModel.create(req.params.serial, serial, req.body, res.locals.issuer));
 
     } catch (error) {
       next(error);
@@ -79,44 +74,44 @@ export const controller = {
   // -------------------------------------------------------------------------------------------------
   async update(req: RecordsApi.UpdateReq, res: RecordsApi.UpdateRes, next: NextFunction) {
     try {
-      const ds: DataStore | null = await dataStoresModel.getBySerial(req.params.serial);
-      
+      const ds: DataStore | null = await dataStoresModel.getBySerial(req.params.serial, { settings: 1 });
+
       if (!ds)
         throw new HttpError(HttpCode.NOT_FOUND, 'dataStoreNotFound');
-  
-      const record = await dataRecordsModel.getBySerial(req.params.serial, req.params.record);
+
+      const record = await dataRecordsModel.getBySerial(req.params.serial, req.params.record, req.body.draft ? DataRecordState.DRAFT : DataRecordState.PUBLISHED);
       const imgsToRemove: string[] = [];
-  
+
       if (!record)
         throw new HttpError(HttpCode.NOT_FOUND, 'recordNotFound');
-  
+
       const files = req.files as Express.Multer.File[] | undefined;
-  
+
       if (files) {
         const paths: Record<string, string> = {};
 
         for (const file of files) {
           const tmpPath = path.join(config.uploadsDir, 'tmp');
           const destPath = path.join(config.uploadsDir, 'fields', req.params.serial, req.params.record);
-          
+
           await fs.copy(tmpPath, destPath);
           paths[file.fieldname] = `/uploads/fields/${req.params.serial}/${req.params.record}/${file.filename}`;
 
           await fs.remove(path.join(tmpPath, file.filename));
-          
+
           imgsToRemove.push(file.fieldname);
         }
 
         Object.assign(req.body, paths);
       }
-  
-      res.json(await dataRecordsModel.update(ds, req.params.record, req.body, res.locals.issuer.serial));
-  
+
+      res.json(await dataRecordsModel.update(req.params.serial, req.params.record, req.body, res.locals.issuer.serial));
+
       // if data store doesnot support history then remove changed images
       if (!ds.settings.history) {
         for (const field of imgsToRemove) {
           const image = record[field];
-    
+
           if (image) {
             const filename = image.slice(image.lastIndexOf('/') + 1);
             await fs.unlink(path.join(config.uploadsDir, 'fields', req.params.serial, req.params.record, filename));
@@ -140,31 +135,35 @@ export const controller = {
   async delete(req: RecordsApi.DeleteReq, res: RecordsApi.DeleteRes, next: NextFunction) {
     try {
       const ds: DataStore | null = await dataStoresModel.getBySerial(req.params.serial, { type: 1, owner: 1, fields: 1, settings: 1 });
-  
+      const isDraft = !!+req.params.draft;
+
       if (!ds)
         throw new HttpError(HttpCode.NOT_FOUND, 'dataStoreNotFound');
-  
-      const record = await dataRecordsModel.getBySerial(req.params.serial, req.params.record);
-  
+
+      const record = await dataRecordsModel.getBySerial(req.params.serial, req.params.record, isDraft ? DataRecordState.DRAFT : DataRecordState.PUBLISHED);
+
       if (!record)
         return res.json(true);
-  
+
       const fields = await dataStoresModel.getFields(req.params.serial);
+
+      res.json(await dataRecordsModel.delete(req.params.serial, req.params.record, isDraft));
+
+      if (!isDraft && typeof ds.settings.workflow.delete !== 'string') {
+        // remove images
+        for (const f of fields.filter((f: Field) => f.type === 'image')) {
+          const image = record[f.name];
   
-      res.json(await dataRecordsModel.delete(ds, req.params.record, res.locals.issuer));
-  
-      for (const f of fields.filter((f: Field) => f.type === 'image')) {
-        const image = record[f.name];
-  
-        if (image) {
-          const filename = image.slice(image.lastIndexOf('/') + 1);
-          await fs.remove(path.join(config.uploadsDir, 'fields', req.params.serial, filename));
+          if (image) {
+            const filename = image.slice(image.lastIndexOf('/') + 1);
+            await fs.remove(path.join(config.uploadsDir, 'fields', req.params.serial, filename));
+          }
         }
-      }
   
-      // delete attachments
-      fs.unlinkSync(path.join(config.uploadsDir, 'attachments', req.params.serial, req.params.record));
-      
+        // delete attachments
+        fs.unlinkSync(path.join(config.uploadsDir, 'attachments', req.params.serial, req.params.record));
+      }
+
     } catch (error) {
       next(error);
     }
